requirements.txt
tensorflow==2.12.0
pandas 
gdown
dvc
mlflow==3.1.1
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .